first functional Dart implementation

+++ lib/app/data/providers/post_provider.dart
@@ -1,37 +1,60 @@
 import 'package:crud/app/data/models/posts_entity.dart';
 import 'package:crud/app/data/repository/post_repository.dart';
+import 'package:dartz/dartz.dart';
 import 'package:flutter/material.dart';
 
 class PostProvider with ChangeNotifier {
   PostProvider({required this.repository});
   PostRepository repository; // = PostRepository();
 
-  List<Post> get posts => repository.fetchPosts();
+  Either<Exception, List<Post>> get posts {
+    var request = repository.fetchPosts();
+    return request.fold((l) => Left(l), (r) => Right(r));
+  }
 
-  int get count => repository.fetchPosts().length;
+  Either<Exception, int> get count {
+    var request = repository.fetchPosts();
+    return request.fold((l) => Left(l), (r) => Right(r.length));
+  }
 
-  void addPost(Post post) {
-    repository.addPost(post);
+  Either<Exception, Post> addPost(Post post) {
+    var request = repository.addPost(post);
+    Either<Exception, Post> response = request.fold((l) => Left(l), (r) => Right(r));
     notifyListeners();
+    return response;
   }
 
-  void updatePost(int index, Post post) {
-    repository.updatePost(index, post);
+  Either<Exception, Post> updatePost(int index, Post post) {
+    var request = repository.updatePost(index, post);
+    Either<Exception, Post> response = request.fold((l) => Left(l), (r) => Right(r));
     notifyListeners();
+    return response;
   }
 
-  void deletePost(int index) {
-    repository.deletePost(index);
+  Either<Exception, int> deletePost(int index) {
+    var request = repository.deletePost(index);
+    Either<Exception, int> response = request.fold((l) => Left(l), (r) => Right(r));
     notifyListeners();
+    return response;
   }
 
-  List<Post> getByUser(int id) {
-    return repository.fetchPostsByUser(id);
+  Either<Exception, List<Post>> getByUser(int id) {
+    var request = repository.fetchPostsByUser(id);
+    Either<Exception, List<Post>> response = request.fold((l) => Left(l), (r) => Right(r));
+    // notifyListeners();
+    return response;
   }
 
-  void deleteByUser(int id) {
-    for (Post p in getByUser(id)) {
-      repository.deletePost(p.id!);
+  Either<Exception, List<Post>> deleteByUser(int id) {
+    var fetchPosts = getByUser(id);
+    var posts = fetchPosts.fold((l) => l, (r) => r);
+    if (posts is Exception) return Left(posts);
+
+    for (Post p in posts as List<Post>) {
+      var request = repository.deletePost(p.id!);
+      var confirmRequest = request.fold((l) => l, (r) => true);
+      if (confirmRequest != true) return Left(confirmRequest as Exception);
     }
+    return Right(posts);
   }
 }
